:recycle: Refactor : Loaded 페이지 리팩토링
타입 보충 및 VAC 패턴 철저히 적용
