+++ src/components/Modal/style.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const DropBack = styled.section`
+	width: 100%;
+	position: fixed;
+	top: 0;
+	bottom: 0;
+	left: 0;
+	right: 0;
+	background-color: rgba(75, 75, 75, 0.6);
+	z-index: 10;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	z-index: 50;
+`;
