+++ src/pages/Loaded/view/LoadedView.tsx
@@ -1,16 +1,27 @@
-import React, { useEffect } from 'react';
-import { Main, Header, Icon, Text } from '../../Loading/style';
+import React from 'react';
+import { Header, Icon, Text } from '../../Loading/style';
 import * as S from '../style';
 import icon from '../../../assets/Food-1.svg';
-import { Props } from '../Loaded';
 import BigWhiteButton from '../../../components/Button/BigWhiteButton/BigWhiteButton';
 import BigRedButton from '../../../components/Button/BigRedButton/BigRedButton';
 import ChainStoreView from './ChainStoreView';
-import { useSelector } from 'react-redux';
-import { RootStateType } from '../../../store';
 import NonChainStore from './NonChainStore';
+import Modal from '../Modal/Modal';
+import { Action } from '../../../slice/findStoreSlice';
 
-export default function LoadedView(props: any) {
+export interface LoadedProps {
+	loadedHeader: string;
+	loadedMessage: string;
+	onClickReturn(): void;
+	onClickModal(): void;
+	renderModal(): boolean;
+	onClickPopUp(): void;
+	data: Action;
+	isNonChainStore(): boolean;
+	isNonData(): boolean;
+}
+
+export default function LoadedView(props: LoadedProps) {
 	return (
 		<S.Main>
 			<Header>
@@ -19,23 +30,19 @@ export default function LoadedView(props: any) {
 			</Header>
 			<Text>{props.loadedMessage}</Text>
 			<S.LoadCont>
-				{props.data?.store_address ? (
-					<S.NonChainImg src={props.data?.file_path + props.data?.file_masking_name} />
-				) : props.data?.review ? (
-					<S.ChainImg src={props.data?.file_path + props.data?.file_masking_name} />
+				{props.isNonChainStore() ? (
+					<S.NonChainImg src={props.data.file_path + props.data.file_masking_name} />
+				) : props.isNonData() ? (
+					<S.ChainImg src={props.data.file_path + props.data.file_masking_name} />
 				) : null}
 
-				{props.data?.store_address ? (
-					<NonChainStore storeData={props.data} />
-				) : (
-					<ChainStoreView storeData={props.data} />
-				)}
+				{props.isNonChainStore() ? <NonChainStore storeData={props.data} /> : <ChainStoreView storeData={props.data} />}
 			</S.LoadCont>
 			<S.ButtonLayout>
 				<BigWhiteButton text="다시 고르기" onclick={props.onClickReturn} />
 				<BigRedButton text="완료하기" onclick={props.onClickModal} />
 			</S.ButtonLayout>
-			{props.renderModal()}
+			{props.renderModal() ? <Modal onClickModal={props.onClickModal} onClickPopUp={props.onClickPopUp} /> : null}
 		</S.Main>
 	);
 }
